ml-service/requirements.txt
flask==3.0.3
flask-cors==4.0.1
scikit-learn
nltk
PyPDF2==3.0.1
python-docx==1.1.2
gunicorn